requirements.txt
pytesseract
discord.py
requests
PIL
numpy